+++ emp-tool/circuits/batcher.h
