+++ emp-tool/circuits/bit.h
