+++ emp-tool/circuits/bit.hpp
