+++ emp-tool/circuits/integer.h
