+++ emp-tool/circuits/integer.hpp
