+++ emp-tool/circuits/number.h
