+++ emp-tool/circuits/swappable.h
