+++ emp-tool/garble/block.h
