+++ emp-tool/garble/garble.h
