+++ emp-tool/garble/garble_gate_halfgates.h
