+++ emp-tool/garble/garble_gate_standard.h
