+++ emp-tool/gc/circuit_file_generator.h
