+++ emp-tool/gc/halfgate_eva.h
