+++ emp-tool/gc/halfgate_gen.h
