+++ emp-tool/io/hash_io_channel.h
