+++ emp-tool/io/net_io_channel.cpp
