+++ emp-tool/utils/block.h
