+++ emp-tool/utils/com.h
