+++ emp-tool/utils/hash.h
