+++ emp-tool/utils/prp.h
