+++ emp-tool/utils/utils.h
