+++ emp-tool/utils/utils.hpp
