+++ emp-tool/utils/utils_ec.h
