+++ emp-tool/utils/utils_ec.hpp
