+++ emp-tool/utils/xor_tree.h
