+++ test/emp-tool-test-bits.vcxproj
@@ -84,7 +84,7 @@
       <WarningLevel>Level3</WarningLevel>
       <Optimization>Disabled</Optimization>
       <SDLCheck>true</SDLCheck>
-      <AdditionalIncludeDirectories>$(ProjectDir)../..;$(ProjectDir)../../../relic/include;$(ProjectDir)../../../libOTe/cryptoTools;C:/libs/boost;C:/libs/mpir/lib/x64/$(configuration)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir)..;$(ProjectDir)../../../relic/include;$(ProjectDir)../../../libOTe/cryptoTools;C:/libs/boost;C:/libs/mpir/lib/x64/$(configuration)</AdditionalIncludeDirectories>
       <BufferSecurityCheck>false</BufferSecurityCheck>
       <EnableParallelCodeGeneration>true</EnableParallelCodeGeneration>
       <AdditionalOptions>/FS %(AdditionalOptions)</AdditionalOptions>
@@ -116,7 +116,7 @@
       <FunctionLevelLinking>true</FunctionLevelLinking>
       <IntrinsicFunctions>true</IntrinsicFunctions>
       <SDLCheck>true</SDLCheck>
-      <AdditionalIncludeDirectories>$(ProjectDir)../..;$(ProjectDir)../../../relic/include;$(ProjectDir)../../../libOTe/cryptoTools;C:/libs/boost;C:/libs/mpir/lib/x64/$(configuration)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir)..;$(ProjectDir)../../../relic/include;$(ProjectDir)../../../libOTe/cryptoTools;C:/libs/boost;C:/libs/mpir/lib/x64/$(configuration)</AdditionalIncludeDirectories>
       <BufferSecurityCheck>false</BufferSecurityCheck>
       <EnableParallelCodeGeneration>true</EnableParallelCodeGeneration>
       <AdditionalOptions>/FS %(AdditionalOptions)</AdditionalOptions>
