+++ test/emp-tool-test-gen_circuit.vcxproj
@@ -84,7 +84,7 @@
       <WarningLevel>Level3</WarningLevel>
       <Optimization>Disabled</Optimization>
       <SDLCheck>true</SDLCheck>
-      <AdditionalIncludeDirectories>$(ProjectDir)../..;$(ProjectDir)../../../relic/include;$(ProjectDir)../../../libOTe/cryptoTools;C:/libs/boost;C:/libs/mpir/lib/x64/$(configuration)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir)../;$(ProjectDir)../../../relic/include;$(ProjectDir)../../../libOTe/cryptoTools;C:/libs/boost;C:/libs/mpir/lib/x64/$(configuration)</AdditionalIncludeDirectories>
       <PreprocessorDefinitions>_WIN32_WINNT=0x0501;THREADING;_CRT_SECURE_NO_WARNINGS;_MBCS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <RuntimeLibrary>MultiThreadedDebug</RuntimeLibrary>
       <AdditionalOptions>/FS %(AdditionalOptions)</AdditionalOptions>
@@ -114,7 +114,7 @@
       <FunctionLevelLinking>true</FunctionLevelLinking>
       <IntrinsicFunctions>true</IntrinsicFunctions>
       <SDLCheck>true</SDLCheck>
-      <AdditionalIncludeDirectories>$(ProjectDir)../..;$(ProjectDir)../../../relic/include;$(ProjectDir)../../../libOTe/cryptoTools;C:/libs/boost;C:/libs/mpir/lib/x64/$(configuration)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir)../;$(ProjectDir)../../../relic/include;$(ProjectDir)../../../libOTe/cryptoTools;C:/libs/boost;C:/libs/mpir/lib/x64/$(configuration)</AdditionalIncludeDirectories>
       <PreprocessorDefinitions>_WIN32_WINNT=0x0501;THREADING;_CRT_SECURE_NO_WARNINGS;_MBCS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
       <AdditionalOptions>/FS %(AdditionalOptions)</AdditionalOptions>
